Migrate the landing page to the index page
